+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## **TBD** [ *1.4.1* ] TBD
+
+* Fix charge time/draw time displaying negative on a decrease.
+* Fix Drang/Mida intrinsics showing up in the filters on some languages.
+* Fix redacted seasons showing up in collections filters.
+* Fix not being able to search for sunset weapons (must enabled "Include Sunset Weapons" first).
+
 ## **Weapon Calculations** [ *1.4.0* ] 2023-03-05
 Adds calculations for weapon properties, a new glossary, and bug fixes.